% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bround}
\alias{bround}
\title{Round a number, preserving extra 0's}
\usage{
bround(x, digits = 1)
}
\arguments{
\item{x}{Number to round.}

\item{digits}{Number of digits past the decimal point to keep.}
}
\value{
A vector of character strings.
}
\description{
Round a number, preserving extra 0's.
}
\details{
From \url{https://github.com/kbroman/broman}

#' @author Karl Broman \email{kbroman@gmail.com}


Uses \code{\link[base]{sprintf}} to round a number, keeping extra 0's.
}
\examples{
myround(51.01, 3)
myround(0.199, 2)

}
\seealso{
\code{\link[base]{round}}, \code{\link[base]{sprintf}}
}
\keyword{utilities}
