% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{close_all}
\alias{close_all}
\title{Close all text sinks and graphics devices}
\usage{
close_all()
}
\value{
none
}
\description{
Close all graphics devices if any are open, issuing a warning.
From \url{https://github.com/pascal-niklaus/pascal/}
}
\author{
Pascal Niklaus \email{pascal.niklaus@ieu.uzh.ch}
}
