% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{grepv}
\alias{grepv}
\alias{greplv}
\alias{subv}
\alias{gsubv}
\title{Search and replace in a vector using a vector of patterns}
\usage{
grepv(x, y, ...)

greplv(x, y, ...)

subv(x, y, ...)

gsubv(x, y, ...)
}
\arguments{
\item{x}{a vector, such as a vector of patterns that you want to search for in a vector}

\item{y}{a vector, the one that you want to search in}

\item{...}{other arguments for \code{\link[base]{grep}}}
}
\value{
a vector
}
\description{
This is a small variation on the standard methods for searching and replacing on matches in a vector, the grep/grepl/sub/gsub functions. Those functions do not natively handle a vector of patterns for matching. This function allows you to convienently pass a vector of patterns to search for matches in a vector.
}
\examples{

x <- letters[1:3]
y <- letters[1:10]

grepv(x, y)

# compare with
# grep(x, y)


}
\author{
Ben Marwick \email{benmarwick@gmail.com}
}
