% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{human_numbers}
\alias{human_numbers}
\title{Human Numbers}
\usage{
human_numbers(x = NULL, smbl = "")
}
\arguments{
\item{x}{a numeric vector to format}

\item{smbl}{a symbol you'd like to prefix your numbers by}
}
\value{
a character vector the same length as the input vector
}
\description{
Format numbers so they're legible for humans. Use this in ggplot for labels where you might use the comma or percent functions from the scales package.
}
\details{
Checks whether numbers are positive or negative.
Allows up to 1 significant figure
sapply used for element-wise application of the humanity function as a vector may include
numbers where billions, millions or thousands are appropriate.

From \url{https://github.com/fdryan}
}
\examples{
human_numbers(c(1000000 , 1500000, 10000000000))
human_numbers(c(1.200000e+05, -2.154660e+05, 2.387790e+05, 4.343500e+04 ,5.648675e+12), "$")
ggplot2 + scale_y_continuous(labels = human_numbers)
ggplot2 + scale_x_continuous(labels = human_numbers)
ggplot2 + scale_x_continuous(labels = human_gbp)
# custom function for large currency values:
human_usd   <- function(x){human_numbers(x, smbl = "$")}
}
