% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{object_sizes}
\alias{object_sizes}
\title{List objects by size}
\usage{
object_sizes(..., n = 10)
}
\arguments{
\item{n}{Number of rows to show}

\item{pos}{.}

\item{pattern}{.}

\item{order.by}{.}

\item{decreasing}{sort by descending?}

\item{head}{.}
}
\value{
a data frame
}
\description{
Lists objects in memory in the current session, and shows how much memory they're using. From \url{http://stackoverflow.com/q/1358003/1036500}
}
\examples{
x <- 1e5
object_sizes()

}
