% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{pull}
\alias{pull}
\title{pull a vector or scalar (vector of length one) from a data frame}
\usage{
pull(x, y)
}
\arguments{
\item{x}{a data frame}

\item{y}{a column name}
}
\value{
a vector or scalar
}
\description{
Sometimes we want to get a single value from a dplyr pipeline so we can use it with inline R code in R markdown. This function is an efficient method of extracting a column from a tibble as a vector. From \url{http://stackoverflow.com/a/24730843/1036500}
}
\examples{

pull(iris, Species)

}

