% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{theme_rotate_x_text}
\alias{theme_rotate_x_text}
\title{theme_rotate_x_text}
\usage{
theme_rotate_x_text(angle = -90, hjust = 0, vjust = 0.25)
}
\arguments{
\item{vjust}{}
}
\description{
Shortcut for rotating the x-axis tick mark labels, because I find it hard to remember the full code for it.
}
